Centralize the default setting overrides in a setting connector in the application extension
