+++ packages/application-extension/package.json
@@ -50,9 +50,11 @@
     "@jupyterlab/mainmenu": "~4.7.0-alpha.0",
     "@jupyterlab/rendermime": "~4.7.0-alpha.0",
     "@jupyterlab/settingregistry": "~4.7.0-alpha.0",
+    "@jupyterlab/statedb": "~4.7.0-alpha.0",
     "@jupyterlab/translation": "~4.7.0-alpha.0",
     "@lumino/coreutils": "^2.2.2",
     "@lumino/disposable": "^2.1.5",
+    "@lumino/polling": "^2.1.5",
     "@lumino/widgets": "^2.7.2"
   },
   "devDependencies": {
