+++ packages/notebook-extension/package.json
@@ -46,7 +46,6 @@
     "@jupyterlab/docmanager": "~4.7.0-alpha.0",
     "@jupyterlab/notebook": "~4.7.0-alpha.0",
     "@jupyterlab/settingregistry": "~4.7.0-alpha.0",
-    "@jupyterlab/statedb": "~4.7.0-alpha.0",
     "@jupyterlab/toc": "~6.7.0-alpha.0",
     "@jupyterlab/translation": "~4.7.0-alpha.0",
     "@lumino/algorithm": "2.0.4",
