+++ packages/tree-extension/src/index.ts
@@ -51,11 +51,6 @@ import { FilesActionButtons } from './fileactions';
 const FILE_BROWSER_FACTORY = 'FileBrowser';
 
 /**
- * The file browser plugin id.
- */
-const FILE_BROWSER_PLUGIN_ID = '@jupyterlab/filebrowser-extension:browser';
-
-/**
  * The namespace for command IDs.
  */
 namespace CommandIDs {
@@ -165,73 +160,6 @@ const fileActions: JupyterFrontEndPlugin<void> = {
     });
     for (const widget of fileActions.widgets) {
       toolbarRegistry.addFactory(FILE_BROWSER_FACTORY, widget.id, () => widget);
-    }
-  },
-};
-
-/**
- * A plugin to set the default file browser settings.
- */
-const fileBrowserSettings: JupyterFrontEndPlugin<void> = {
-  id: '@jupyter-notebook/tree-extension:settings',
-  description: 'Set up the default file browser settings',
-  requires: [IDefaultFileBrowser],
-  optional: [ISettingRegistry],
-  autoStart: true,
-  activate: (
-    app: JupyterFrontEnd,
-    browser: IDefaultFileBrowser,
-    settingRegistry: ISettingRegistry | null
-  ) => {
-    // Default config for notebook.
-    // This is a different set of defaults than JupyterLab.
-    const defaultFileBrowserConfig = {
-      navigateToCurrentDirectory: false,
-      singleClickNavigation: true,
-      showLastModifiedColumn: true,
-      showDateCreatedColumn: false,
-      showFileSizeColumn: true,
-      showHiddenFiles: false,
-      showFileCheckboxes: true,
-      sortNotebooksFirst: true,
-      sortFileNamesNaturally: true,
-      showFullPath: false,
-      clearFilterOnNavigation: true,
-      allowFileUploads: true,
-    };
-
-    // Apply defaults on plugin activation
-    let key: keyof typeof defaultFileBrowserConfig;
-    for (key in defaultFileBrowserConfig) {
-      const value = defaultFileBrowserConfig[key];
-      // only set the value if it is different, since some setters trigger
-      // a relayout of the file browser
-      if (browser[key] !== value) {
-        browser[key] = value;
-      }
-    }
-
-    // Refresh the file browser after applying the defaults, which also
-    // resets the refresh poll in case an earlier tick failed and left it
-    // backing off
-    void browser.model.refresh();
-
-    if (settingRegistry) {
-      void settingRegistry.load(FILE_BROWSER_PLUGIN_ID).then((settings) => {
-        function onSettingsChanged(settings: ISettingRegistry.ISettings): void {
-          let key: keyof typeof defaultFileBrowserConfig;
-          for (key in defaultFileBrowserConfig) {
-            const value = settings.get(key).user as boolean;
-            // only set the setting if it is defined by the user and different,
-            // since some setters trigger a relayout of the file browser
-            if (value !== undefined && browser[key] !== value) {
-              browser[key] = value;
-            }
-          }
-        }
-        settings.changed.connect(onSettingsChanged);
-        onSettingsChanged(settings);
-      });
     }
   },
 };
@@ -458,7 +386,6 @@ const notebookTreeWidget: JupyterFrontEndPlugin<INotebookTree> = {
 const plugins: JupyterFrontEndPlugin<any>[] = [
   createNew,
   fileActions,
-  fileBrowserSettings,
   fileFilterCommand,
   loadPlugins,
   openFileBrowser,
